o Add support for variable-length tags (no I/O yet)
o Add API to get/set tag values as lists of pointers to per-entity values.

o Fix handling of default value for tags on mesh.


New APIs:

tag_create_variable_length(..) 
  - need extra argument for length of default value
  - old tag_create w/ size == MB_VARIABLE_LENGTH can also be used if no 
    default value.

tag_get_default_value( MBTag, const void*&, int& size ) const
  - get reference to internal storage for tag default value
  - passes back size of tag value also
  - old tag_get_default_value cannot be used for variable-length tags
  - new tag_get_default_value can be used for fixed- or variable-length tags

tag_get_data( MBTag, const MBEntityHandle*, int, const void** data, int* sizes = 0 )
tag_get_data( MBTag, const MBRange&,             const void** data, int* sizes = 0 )
  - get pointers to internal storage for each tag value
  - get size (in bytes) for each tag value
  - 'sizes' is optional for fixed-length tags, but will be set appropriately 
    if passed.
  - 'sizes' is mandatory for variable-length tags

tag_set_data( MBTag, const MBEntityHandle*, int, void const* const* data, int* sizes = 0 )
tag_set_data( MBTag, const MBRange&,             void const* const* data, int* sizes = 0 )
  - set tag data given separate pointers to each tag value
  - 'sizes' is optional for fixed-length tags, but will be set appropriately 
     if passed.
  - 'sizes' is mandatory for variable-length tags
  - if 'sizes' is specified for fixed-length tags, it must contain the correct
    values (same value for every entity.)


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1576 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ SequenceData.cpp
@@ -153,7 +153,7 @@ void SequenceData::move_tag_data( SequenceData* destination, TagServer* tag_serv
 
 void SequenceData::release_tag_data( const int* tag_sizes, int num_tag_sizes )
 {
-  assert( num_tag_sizes >= numTagData );
+  assert( num_tag_sizes >= (int)numTagData );
   for (unsigned i = 1; i <= numTagData; ++i)
     release_tag_data( i, tag_sizes[i-1] );
 }
